+++ src/Http/Controllers/ResourceController.php
@@ -68,6 +68,9 @@ class ResourceController extends Controller
      */
     public function create($module)
     {
+        if(!auth()->user()->can('create-'.$module) && !auth()->user()->hasAnyRole('SuperAdmin|Admin')){
+            abort(403);
+        }
         $model = Module::where('title', $module)->first();
         if(null == $options =json_decode($model->options)){
             $options = [];
@@ -112,6 +115,9 @@ class ResourceController extends Controller
      */
     public function store(ResourceRequest $request,$module)
     {
+        if(!auth()->user()->can('create-'.$module) && !auth()->user()->hasAnyRole('SuperAdmin|Admin')){
+            abort(403);
+        }
 
 //        dd($request->all(), $module);
 //        dd($this->def_lang);
@@ -151,6 +157,9 @@ class ResourceController extends Controller
      */
     public function edit($module, $id)
     {
+        if(!auth()->user()->can('edit-'.$module) && !auth()->user()->hasAnyRole('SuperAdmin|Admin')){
+            abort(403);
+        }
         $model = Module::where('title', $module)->first();
         if(null == $options =json_decode($model->options)){
             $options = [];
@@ -210,6 +219,9 @@ class ResourceController extends Controller
     public function update(ResourceRequest $request, $module,  $id)
     {
 //        dd($request->all());
+        if(!auth()->user()->can('update-'.$module) && !auth()->user()->hasAnyRole('SuperAdmin|Admin')){
+            abort(403);
+        }
 
         if($resource =  $this->CRUD->update($id, $request->all())) {
             if ($request->has('meta')) {
@@ -232,6 +244,9 @@ class ResourceController extends Controller
      */
     public function destroy($module, $id)
     {
+        if(!auth()->user()->can('delete-'.$module) && !auth()->user()->hasAnyRole('SuperAdmin|Admin')){
+            abort(403);
+        }
         if($this->CRUD->destroy($id)){
             return redirect()->back()->with('success', Str::singular(ucwords($module)).' Successfully Deleted.');
         }
@@ -240,6 +255,9 @@ class ResourceController extends Controller
 
     public function translate($module, $id, $lang )
     {
+        if(!auth()->user()->can('update-'.$module) && !auth()->user()->hasAnyRole('SuperAdmin|Admin')){
+            abort(403);
+        }
         $translate = $this->CRUD->createOrEditResourceTranslation($module, $id, $lang );
         $model = Module::where('title', $module)->first();
